app/config.py
import os

from dotenv import load_dotenv

load_dotenv()

BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))


class Config:
    SECRET_KEY = os.environ.get("FLASK_SECRET", "dev-secret-change-me")
   
    SQLALCHEMY_DATABASE_URI = os.environ.get(
        "DATABASE_URL", f"sqlite:///{os.path.join(BASE_DIR, 'chat_history.db')}"
    )
    SQLALCHEMY_TRACK_MODIFICATIONS = False
